FEAT: Modularized HistoricalFng
